requirements.txt
openai>=0.27.8
python-dotenv>=1.0.0
transformers>=4.31.0